+++ cli/src/main.rs
@@ -7,65 +7,11 @@ extern crate rustyline;
 extern crate dirs;
 
 mod agent;
+mod startup;
 
 use clap::Parser;
-use anyhow::{Context, Result};
-use rustyline::{Editor};
+use anyhow::{Result};
 
-
-#[tokio::main]
-async fn main() -> Result<()> {
-    //let path = "test.txt";
-    //let content = std::fs::read_to_string(path)
-    //    .with_context(|| format!("could not read file `{}`", path))?;
-    //println!("file content: {}", content);
-
-    let home_dir = dirs::home_dir().expect("Could not get home directory");
-    let data_path = home_dir.join(".ad4m-cli");
-    if !data_path.exists() {
-        std::fs::create_dir_all(&data_path).with_context(|| format!("Could not create directory `{}`", data_path.display()))?;
-    }
-
-    let mut cap_token = String::new();
-
-    let cap_token_file = data_path.join("cap_token");
-    if cap_token_file.exists() {
-        cap_token = std::fs::read_to_string(&cap_token_file).with_context(|| format!("Could not read file `{}`", cap_token_file.display()))?;
-        println!("Found cap token in file.");
-    } else {
-        println!("No cap token found in file. Requesting one...");
-
-        if let Ok(request_id) = agent::run_request_capability().await {
-            println!("Got request id: {:#?}", request_id);
-    
-            let mut rl = Editor::<()>::new()?;
-            if let Ok(rand) = rl.readline("Enter random string: ") {
-                let result = agent::run_retrieve_capability(request_id, rand).await;
-                match result {
-                    Ok(jwt) => {
-                        cap_token = jwt.clone();
-                        std::fs::write(&cap_token_file, jwt).with_context(|| format!("Could not write file `{}`", cap_token_file.display()))?;
-                        println!("Wrote cap token to file.");
-                    },
-                    Err(e) => {
-                        eprintln!("Error generating capability token: {:#?}", e);
-                    }
-                }
-            }
-        } else {
-            println!("Error requesting capability");
-        }
-    }
-
-
-
-
-
-
-    println!("cap token: {}", cap_token);
-
-    Ok(())
-}
 /// AD4M command line client
 #[derive(Parser, Debug)]
 #[command(author, version, about, long_about = None)]
@@ -79,12 +25,13 @@ struct Args {
    count: u8,
 }
 
-/*
-fn main() {
-   let args = Args::parse();
+#[tokio::main]
+async fn main() -> Result<()> {
+    //let args = Args::parse();
 
-   for _ in 0..args.count {
-       println!("Hello {}!", args.name)
-   }
+    let cap_token = startup::get_cap_token().await?;
+
+    println!("cap token: {}", cap_token);
+
+    Ok(())
 }
- */
